matching-service/requirements.txt
flask==3.1.1
flask-cors==5.0.1
sentence-transformers==3.4.1
scikit-learn==1.6.1
numpy==2.2.4
python-dotenv==1.1.0
gunicorn==23.0.0
